refactor: TRS-22 패키지 재정렬
