+++ back/tripsok_back/src/main/java/com/tripsok_back/model/theme/Theme.java
@@ -1,4 +1,4 @@
-package com.tripsok_back.model.Theme;
+package com.tripsok_back.model.theme;
 
 import com.tripsok_back.support.BaseTimeEntity;
