+++ back/tripsok_back/src/main/java/com/tripsok_back/repository/ThemeRepository.java
@@ -2,7 +2,7 @@ package com.tripsok_back.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.tripsok_back.model.Theme.Theme;
+import com.tripsok_back.model.theme.Theme;
 
 public interface ThemeRepository extends JpaRepository<Theme, Integer> {
 }
