+++ back/tripsok_back/src/main/java/com/tripsok_back/repository/place/AccommodationRepository.java
@@ -1,4 +1,4 @@
-package com.tripsok_back.batch.repository;
+package com.tripsok_back.repository.place;
 
 import java.util.Optional;
 
@@ -9,9 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.tripsok_back.model.place.Place;
 
 @Repository
-public interface RestaurantRepository extends JpaRepository<Place, Integer> {
-
-	@EntityGraph(attributePaths = {"restaurant"})
+public interface AccommodationRepository extends JpaRepository<Place, Integer> {
+	@EntityGraph(attributePaths = {"accommodation"})
 	Optional<Place> findByContentId(Integer contentId);
-
 }
