+++ back/tripsok_back/src/main/java/com/tripsok_back/repository/place/PlaceRepository.java
@@ -1,4 +1,4 @@
-package com.tripsok_back.batch.repository;
+package com.tripsok_back.repository.place;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
