+++ back/tripsok_back/src/main/java/com/tripsok_back/repository/place/TourRepository.java
@@ -1,4 +1,4 @@
-package com.tripsok_back.batch.repository;
+package com.tripsok_back.repository.place;
 
 import java.util.Optional;
