+++ back/tripsok_back/src/main/java/com/tripsok_back/service/InterestThemeService.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
-import com.tripsok_back.model.Theme.Theme;
+import com.tripsok_back.model.theme.Theme;
 import com.tripsok_back.model.user.InterestTheme;
 import com.tripsok_back.model.user.TripSokUser;
 import com.tripsok_back.repository.InterestThemeRepository;
